args: Move parse_char to args and improve
- Accept 0x format (0x00)
- Raise error if char is empty
- Handle None in parse_char
